Move Strict tests to a new test class.
